backend/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
